feat(bot): introduce user-mode "open door" command
- Close door is an admin-only option
- Bot checks your orders before opening (NOT TESTED)
