+++ vk_bot/src/settings.py
@@ -4,3 +4,5 @@ ADMIN_HARDCODED_LIST = [
     TECHNICAL_ADMIN_ID,
     CHIEF_ID
 ]
+
+UNLOCK_CHECK_URL = 'https://8ka.mipt.ru/services/washing/{service_id}/unlock/'
